Add integration test
