+++ unique_toolkit/pyproject.toml
@@ -32,6 +32,7 @@ pre-commit = "^3.7.1"
 pytest-asyncio = "^0.23.8"
 pytest-mock = "^3.14.0"
 unique-sdk = {path = "../unique_sdk"}
+ipykernel = "^6.29.5"
 
 [build-system]
 requires = ["poetry-core"]
