+++ unique_toolkit/tests/language_model/test_language_models_service_integration.py
@@ -1,8 +1,5 @@
 from datetime import datetime
 
-import pytest
-from pydantic import ValidationError
-
 from unique_toolkit.app.schemas import Event
 from unique_toolkit.content.schemas import ContentChunk, ContentMetadata
 from unique_toolkit.language_model.infos import LanguageModelName
@@ -230,18 +227,3 @@ class TestLanguageModelServiceIntegration:
         assert response.tool_calls is not None
         assert response.tool_calls[0].name == "get_weather"
 
-    def test_language_model_tool_name_pattern(self):
-        with pytest.raises(ValidationError):
-            LanguageModelTool(
-                name="invalid name!",
-                description="Invalid tool name",
-                parameters=LanguageModelToolParameters(
-                    type="object",
-                    properties={
-                        "param": LanguageModelToolParameterProperty(
-                            type="string", description="A parameter"
-                        )
-                    },
-                    required=["param"],
-                ),
-            )
